+++ doc/whats_new.rst
@@ -23,6 +23,12 @@ Changes
  - cmap is now edge_cmap
  - threshold is now edge_threshold
 
+- plotting.view_markers() default marker color is now 'red', consistent with add_markers().
+- plotting.view_markers() parameter names are consistent with add_markers():
+
+ - coords is now marker_coords
+ - colors is now marker_color
+
 0.5.0
 =====